Remove angular.json items from Step 2 of readme
The bootstrap style declaration and bootstrap scripts declaration
were creating errors. Like saying that bootstrap loaded in before
jquery which it depends on. These dependencies were not in the
project before they were put in the markdown, so they are likely
not necessary. If anything, the bootstrap.css was conflicting
with `bootstrap.min.css` in styles that was already there.