Move buttons to header, leave remnants of attempt to make it fixed
I think that the header of the workout page should be fixed over
the content. The `#header` css that is commented out is part of
that process but I couldn't figure that out yet. It will have to
interact with the navbar's css, but I am still relearning how the
nav is rendered in Angular2.